# server/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MotionMesh API
  version: 1.0.0
components:
  schemas:
...
    CDNDomain:
      type: object
      properties:
        id:
          type: string
        account_id:
          type: string
        hostname:
          type: string
        cloudflare_hostname_id:
          type: string
        hostname_status:
          type: string
        ssl_status:
          type: string
        verification_errors:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []
paths:
...
  /v1/billing/portal:
    post:
      summary: Create billing portal session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                return_url:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
  /v1/billing/checkout:
    post:
      summary: Create checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                price_id:
                  type: string
                return_url:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
  /v1/billing/funds:
    post:
      summary: Add funds to balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                  description: Amount in cents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
  /v1/branding:
    get:
      summary: Get active watermark
      responses:
        '200':
          description: OK
    put:
      summary: Update watermark configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: string
                opacity:
                  type: number
      responses:
        '200':
          description: OK
  /v1/branding/asset:
    post:
      summary: Upload watermark asset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                asset:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
  /v1/videos:
    get:
      summary: List videos
      responses:
        '200':
          description: OK
    post:
      summary: Upload video initiation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                size_bytes:
                  type: number
      responses:
        '201':
          description: Created
...
  /v1/cdn/domains:
    get:
      summary: List CDN domains
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CDNDomain'
    post:
      summary: Add custom CDN domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hostname:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CDNDomain'
  /v1/cdn/domains/{id}:
    get:
      summary: Get CDN domain status
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CDNDomain'
    delete:
      summary: Delete CDN domain
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
...